fix: resolved issue in which players could persist in the game even if they left earlier
