+++ src/app/game/game-mode/base-game-mode/setup-state.ts
@@ -1,6 +1,5 @@
 import { VictoryManager } from 'src/app/managers/victory-manager';
 import { PlayerManager } from 'src/app/player/player-manager';
-import { PLAYER_STATUS } from 'src/app/player/status/status-enum';
 import { ScoreboardManager } from 'src/app/scoreboard/scoreboard-manager';
 import { SettingsContext } from 'src/app/settings/settings-context';
 import { GlobalGameData } from '../../state/global-game-state';
@@ -11,7 +10,6 @@ import { Quests } from 'src/app/quests/quests';
 import { clearTickUI } from '../utillity/update-ui';
 import { TeamManager } from 'src/app/teams/team-manager';
 import { TreeManager } from '../../services/tree-service';
-import { ParticipantEntityManager } from 'src/app/utils/participant-entity';
 import { ReplayManager } from 'src/app/statistics/replay-manager';
 import { CountdownMessage } from '../../../utils/messages';
 
@@ -21,7 +19,7 @@ export class SetupState<T extends StateData> extends BaseState<T> {
 	}
 
 	run(): void {
-		CountdownMessage("Initializing the game");
+		CountdownMessage('Initializing the game');
 
 		clearTickUI();
 
@@ -29,44 +27,6 @@ export class SetupState<T extends StateData> extends BaseState<T> {
 
 		SettingsContext.getInstance().applyStrategy('Promode');
 		SettingsContext.getInstance().applyStrategy('Diplomacy');
-
-		// Remove irrelevant players from the game
-		GlobalGameData.matchPlayers.forEach((val) => {
-			val.trackedData.reset();
-			val.trackedData.setKDMaps();
-			if (GetPlayerSlotState(val.getPlayer()) == PLAYER_SLOT_STATE_PLAYING) {
-				val.status.set(PLAYER_STATUS.ALIVE);
-			} else {
-				val.status.set(PLAYER_STATUS.LEFT);
-				PlayerManager.getInstance().players.delete(val.getPlayer());
-			}
-		});
-
-		const participants = ParticipantEntityManager.getParticipantEntities();
-		ParticipantEntityManager.executeByParticipantEntities(
-			participants,
-			(_) => {},
-			(team) => {
-				team.reset();
-			}
-		);
-		TeamManager.getInstance()
-			.getTeams()
-			.forEach((team) => {
-				team.reset();
-			});
-
-		const players = [...PlayerManager.getInstance().players.values()];
-		GlobalGameData.prepareMatchData(players);
-
-		// Prepare stat tracking
-		GlobalGameData.matchPlayers.forEach((player) => {
-			SetPlayerState(player.getPlayer(), PLAYER_STATE_RESOURCE_GOLD, 0);
-			player.status.set(PLAYER_STATUS.ALIVE);
-			player.status.status = PLAYER_STATUS.ALIVE;
-			player.trackedData.bonus.showForPlayer(player.getPlayer());
-			player.trackedData.bonus.repositon();
-		});
 
 		// Setting up the scoreboard
 		if (SettingsContext.getInstance().isFFA() || GlobalGameData.matchPlayers.length <= 2) {
