+++ src/app/game/game-mode/mode/standard-mode.ts
@@ -10,12 +10,14 @@ import { StateData } from '../state/state-data';
 import { VisionState } from '../base-game-mode/vision-state';
 import { ApplyFogState } from '../base-game-mode/apply-fog-state';
 import { DisablePausesState } from '../base-game-mode/disable-pauses-state';
+import { UpdatePlayerStatusState } from '../base-game-mode/update-player-status-state';
 
 export class StandardData implements StateData {}
 
 export class StandardMode extends BaseMode<StandardData> {
 	protected setupStates() {
 		return [
+			new UpdatePlayerStatusState(),
 			new DisablePausesState(),
 			new SetupState(),
 			new ApplyFogState(),
