+++ src/app/game/game-mode/mode/w3c-mode.ts
@@ -16,12 +16,14 @@ import { W3C_TERMINATE_IF_ALONE_HUMAN_PLAYER } from 'src/configs/game-settings';
 import { ApplyFogState } from '../base-game-mode/apply-fog-state';
 import { ProModeGameLoopState } from '../promode-game-mode/promode-game-loop-state';
 import { W3CTipsState } from '../w3c-mode/w3c-tips-state';
+import { UpdatePlayerStatusState } from '../base-game-mode/update-player-status-state';
 
 export class W3CData implements StateData {}
 
 export class W3CMode extends BaseMode<W3CData> {
 	protected setupStates() {
 		const states = [
+			new UpdatePlayerStatusState(),
 			new SetupState(),
 			new ApplyFogState(),
 			new CityDistributeState(),
