+++ src/app/game/game-mode/state/state-data.ts
@@ -1,1 +1,2 @@
-export interface StateData {}
+export interface StateData {
+}
